Add model classes and tests.
