+++ kata/olinqpics/olinqpics.sln
@@ -1,0 +1,28 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "olinqpics", "olinqpics\olinqpics.csproj", "{B65FB69D-426D-4510-807A-9499899BF689}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "olinqpics.tests", "olinqpics.tests\olinqpics.tests.csproj", "{886A1039-D838-4FB6-8935-50FF8428F2DA}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{B65FB69D-426D-4510-807A-9499899BF689}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B65FB69D-426D-4510-807A-9499899BF689}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B65FB69D-426D-4510-807A-9499899BF689}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B65FB69D-426D-4510-807A-9499899BF689}.Release|Any CPU.Build.0 = Release|Any CPU
+		{886A1039-D838-4FB6-8935-50FF8428F2DA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{886A1039-D838-4FB6-8935-50FF8428F2DA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{886A1039-D838-4FB6-8935-50FF8428F2DA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{886A1039-D838-4FB6-8935-50FF8428F2DA}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
