+++ kata/olinqpics/olinqpics/Gender.cs
@@ -1,0 +1,6 @@
+enum Gender
+{
+	Male,
+	Female,
+	NonBinary
+}
