+++ kata/olinqpics/olinqpics/Person.cs
@@ -1,0 +1,7 @@
+class Person 
+{
+	public int Age { get; set; }
+	public string? Name { get; set; }
+	public Gender Gender { get; set; }
+	public IEnumerable<string>? FavouriteColours { get; set; }
+}
